requirements.txt
fastapi==0.115.0
uvicorn==0.22.0
pydantic==1.10.11
SQLAlchemy==1.4.41
mysql-connector-python==8.1.0
Jinja2==3.1.2
aiofiles==23.1.0
python-dotenv==1.0.0
PyMySQL==1.0.2
python-multipart==0.0.5
cryptography==41.0.3